tests/wpt/tests/html/browsers/the-window-object/window-aliases.html

<!doctype html>
<meta charset=utf-8>
<title>Aliases of the window object</title>
<link rel="author" title="Ms2ger" href="mailto:Ms2ger@gmail.com">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-window">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-frames">
<link rel="help" href="https://html.spec.whatwg.org/multipage/#dom-self">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
var global = this;

test(function() {
  assert_equals(window, global);
  assert_equals(window.window, global);
}, "window should be the global object");

test(function() {
  assert_equals(frames, global);
  assert_equals(window.frames, global);
}, "frames should be the global object");

test(function() {
  assert_equals(self, global);
  assert_equals(window.self, global);
}, "self should be the global object");
</script>
